requirements.txt
Django==1.9.5
requests==2.9.1
djangorestframework==3.3.3
